requirements.txt
cycler==0.10.0
future==0.18.2
matplotlib==3.4.1
numpy==1.20.2
opencv-python==4.5.1.48
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.2
seaborn==0.11.0
six==1.15.0
torch==1.6.0
torchaudio==0.6.0
torchvision==0.7.0
PyYAML==5.3.1
pytorch_msssim
